additional grains osqueryversion and osquerybinpath
The osqueryversion grain would make it possible to match based osquery
version found on the host. This would allow specific query profiles
to be matched only where new enough version is present.

The osquerybinpath grain returns the absolute path of the osqueryi binary.
In some environments the PATH environemnt can't be depended on to include
the location of the binaries. Having the location of the binary resolved
once at the start would lessen chances for surprises later on.