There probably isn't ever a good time to tell Splunk to index extra stuff.
The changes to the indexed fields intended by the removed code have to be
accompanied by changes to the fields configuration in Splunk or oddball
problems will occur (e.g., duplicate field indexing). The best plan (if you
really really need extra index fields) is to ask the Splunk admins to add them
to the index through their usual methods.

This affords them the opportunity to say, "that's not a good idea because ____"
and prevents the data from getting wonky in their indexes.

There are certainly fields one might wish to add to the index, it's best to
coordinate them carefully with the Splunk admins. (And these fields are more
rare than you'd think if you're accustomed to traditional databases.)