Don't strip empty dict entries in HEC code
The stripping of empty fields already happens in the returners
themselves. Additionally, doing this recursively was going overboard --
if I have have nested data with None values, I want to know.

If we do need to bring this back eventually, we need to use
copy.deepcopy or similar to avoid side effects. This was hard to track
down.