Remove splunkconfig grains
As far as I can tell, this doesn't work, and I'm not sure it ever did.
We do get splunk config in grains, but the splunk returners don't load
out of grains unless the splunk config is missing in the main config
(which it never should be).

Also, this results in secrets in grains which I don't like.