bug fix: filtering opts that have dots in their name
opts in the patterns_to_filter that had a dot in them weren't filtered
because the dot was replaced with underscore before _remove_sensitive_info
was called. this could have been fixed also by simply replacing the dot
with the underscore in the patterns_to_filter list but the list would have
been less intuitive to operate.