Remove StartLimit directives in systemd unit files
This change has two driving reasons. The first is that the StartLimit
directives have changed in recent versions of systemd. In some versions
it's expected in the `[Service]` section, in others in the `[Unit]`
section. Additionally, it's also been renamed to `StartLimitInvervalSec`
in recent versions of systemd.

Rather than deal with these compatibility issues, I propose that we
remove them. I have also bumped the `RestartSec` to 300. Basically, if
hubble ever crashes, we do want it to restart, but we're not in a hurry.
So I think using `RestartSec=300`, the impact of trying to restart
hubble forever is negligible, and we save ourselves the hassle of trying
to deal with systemd compatibility.